Generate new Rails application
A new Rails application has been generated using the following command:

    rails new -d postgresql --webpack=stimulus

PostgreSQL is our database of choice due to its stability and features.
To ensure that the development and production environment are as similar
as possible, Postgres is used in all three environments.

Stimulus is used as a lightweight JavaScript framework to organize and
structure the code. It has been automatically added to the project
through Webpack.