Fix query for weekly activities
The query for the weekly activities did not work correctly. Adding the
`where` clause caused habits without activities in the current week, but
activities in another week, to disappear. Different variations of
`includes` and `left_outer_joins` were tested without success.

The current implementation is suboptimal in terms of efficiency and
database queries. Some of this is caught by ActiveRecord's caching
layer, but in the long term this part of the app should be improved.